Rename new_then_execute() to then_execute()
