+++ agency/detail/executor_traits/multi_agent_then_execute_returning_default_container.hpp
@@ -53,7 +53,7 @@ typename new_executor_traits<Executor>::template future<
     >
   >;
 
-  return new_executor_traits<Executor>::new_then_execute(ex, f, container_factory<container_type>{}, shape, fut);
+  return new_executor_traits<Executor>::then_execute(ex, f, container_factory<container_type>{}, shape, fut);
 } // end multi_agent_then_execute_returning_default_container()
