+++ agency/detail/executor_traits/multi_agent_then_execute_returning_user_specified_container.hpp
@@ -24,23 +24,25 @@ struct use_multi_agent_then_execute_returning_user_specified_container_member_fu
 
 struct use_multi_agent_then_execute_with_shared_inits_returning_user_specified_container {};
 
-template<class Container, class Executor, class Function, class Future>
+template<class Executor, class Function, class Factory, class Future>
 using select_multi_agent_then_execute_returning_user_specified_container_implementation =
   typename std::conditional<
-    has_multi_agent_then_execute_returning_user_specified_container<Container,Executor,Function,Future>::value,
+    has_multi_agent_then_execute_returning_user_specified_container<Executor,Function,Factory,Future>::value,
     use_multi_agent_then_execute_returning_user_specified_container_member_function,
     use_multi_agent_then_execute_with_shared_inits_returning_user_specified_container
   >::type;
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class Future>
+template<class Executor, class Function, class Factory, class Future>
 __AGENCY_ANNOTATION
-typename new_executor_traits<Executor>::template future<Container>
+typename new_executor_traits<Executor>::template future<
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+>
   multi_agent_then_execute_returning_user_specified_container(use_multi_agent_then_execute_returning_user_specified_container_member_function,
-                                                              Executor& ex, Function f, typename new_executor_traits<Executor>::shape_type shape, Future& fut)
+                                                              Executor& ex, Function f, Factory result_factory, typename new_executor_traits<Executor>::shape_type shape, Future& fut)
 {
-  return ex.template then_execute<Container>(f, shape, fut);
+  return ex.then_execute(f, result_factory, shape, fut);
 } // end multi_agent_then_execute_returning_user_specified_container()
 
 
@@ -75,28 +77,32 @@ struct ignore_tail_parameters_and_invoke<Function,void>
 };
 
 
-template<class Container, size_t... Indices, class Executor, class Function, class Future, class Tuple>
+template<size_t... Indices, class Executor, class Function, class Factory, class Future, class Tuple>
 __AGENCY_ANNOTATION
-typename new_executor_traits<Executor>::template future<Container>
+typename new_executor_traits<Executor>::template future<
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+>
   multi_agent_then_execute_returning_user_specified_container_impl(detail::index_sequence<Indices...>,
-                                                                   Executor& ex, Function f, typename new_executor_traits<Executor>::shape_type shape, Future& fut,
+                                                                   Executor& ex, Function f, Factory result_factory, typename new_executor_traits<Executor>::shape_type shape, Future& fut,
                                                                    const Tuple& tuple_of_unit_factories)
 {
   using value_type = typename future_traits<Future>::value_type;
 
-  return new_executor_traits<Executor>::template then_execute<Container>(ex, ignore_tail_parameters_and_invoke<Function,value_type>{f}, shape, fut, std::get<Indices>(tuple_of_unit_factories)...);
+  return new_executor_traits<Executor>::then_execute(ex, ignore_tail_parameters_and_invoke<Function,value_type>{f}, result_factory, shape, fut, std::get<Indices>(tuple_of_unit_factories)...);
 } // end multi_agent_then_execute_returning_user_specified_container_impl()
 
 
-template<class Container, class Executor, class Function, class Future>
+template<class Executor, class Function, class Factory, class Future>
 __AGENCY_ANNOTATION
-typename new_executor_traits<Executor>::template future<Container>
+typename new_executor_traits<Executor>::template future<
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+>
   multi_agent_then_execute_returning_user_specified_container(use_multi_agent_then_execute_with_shared_inits_returning_user_specified_container,
-                                                              Executor& ex, Function f, typename new_executor_traits<Executor>::shape_type shape, Future& fut)
+                                                              Executor& ex, Function f, Factory result_factory, typename new_executor_traits<Executor>::shape_type shape, Future& fut)
 {
   auto tuple_of_unit_factories = new_executor_traits_detail::make_tuple_of_unit_factories(ex);
 
-  return multi_agent_then_execute_returning_user_specified_container_impl<Container>(detail::make_index_sequence<std::tuple_size<decltype(tuple_of_unit_factories)>::value>(), ex, f, shape, fut, tuple_of_unit_factories);
+  return multi_agent_then_execute_returning_user_specified_container_impl(detail::make_index_sequence<std::tuple_size<decltype(tuple_of_unit_factories)>::value>(), ex, f, result_factory, shape, fut, tuple_of_unit_factories);
 } // end multi_agent_then_execute_returning_user_specified_container()
 
 
@@ -106,24 +112,26 @@ typename new_executor_traits<Executor>::template future<Container>
 
 
 template<class Executor>
-  template<class Container, class Function, class Future,
+  template<class Function, class Future, class Factory,
            class Enable1,
-           class Enable2,
-           class Enable3
+           class Enable2
            >
 __AGENCY_ANNOTATION
-typename new_executor_traits<Executor>::template future<Container>
+typename new_executor_traits<Executor>::template future<
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+>
   new_executor_traits<Executor>
     ::then_execute(typename new_executor_traits<Executor>::executor_type& ex,
                    Function f,
+                   Factory result_factory,
                    typename new_executor_traits<Executor>::shape_type shape,
                    Future& fut)
 {
   namespace ns = detail::new_executor_traits_detail::multi_agent_then_execute_returning_user_specified_container_implementation_strategies;
 
-  using check_for_member_function = ns::select_multi_agent_then_execute_returning_user_specified_container_implementation<Container,Executor,Function,Future>;
+  using check_for_member_function = ns::select_multi_agent_then_execute_returning_user_specified_container_implementation<Executor,Function,Factory,Future>;
 
-  return ns::multi_agent_then_execute_returning_user_specified_container<Container>(check_for_member_function(), ex, f, shape, fut);
+  return ns::multi_agent_then_execute_returning_user_specified_container(check_for_member_function(), ex, f, result_factory, shape, fut);
 } // end new_executor_traits::then_execute()
