+++ agency/detail/executor_traits/multi_agent_then_execute_with_shared_inits_returning_user_specified_container.hpp
@@ -24,23 +24,23 @@ struct use_multi_agent_then_execute_with_shared_inits_returning_user_specified_c
 
 using use_strategy_1 = use_multi_agent_then_execute_with_shared_inits_returning_user_specified_container_member_function;
 
-template<class Executor, class Container, class Function, class Future, class... Factories>
+template<class Executor, class Function, class Factory, class Future, class... Factories>
 using has_strategy_1 = has_multi_agent_then_execute_with_shared_inits_returning_user_specified_container<
-  Container,
   Executor,
   Function,
+  Factory,
   Future,
   Factories...
 >;
 
 
-template<class Executor, class Container, class Function, class Future, class... Types>
+template<class Executor, class Function, class Factory, class Future, class... Types>
 struct has_strategy_1_workaround_nvbug_1665745
 {
   using type = has_multi_agent_then_execute_with_shared_inits_returning_user_specified_container<
-    Container,
     Executor,
     Function,
+    Factory,
     Future,
     Types...
   >;
@@ -52,13 +52,17 @@ struct use_multi_agent_when_all_execute_and_select_with_shared_inits_member_func
 
 using use_strategy_2 = use_multi_agent_when_all_execute_and_select_with_shared_inits_member_function;
 
-template<class Executor, class Container, class Function, class Future, class... Factories>
+template<class Executor, class Function, class Factory, class Future, class... Factories>
 using has_strategy_2 = has_multi_agent_when_all_execute_and_select_with_shared_inits<
   detail::index_sequence<0>,
   Executor,
   test_function_returning_void,
   detail::tuple<
-    typename new_executor_traits<Executor>::template future<Container>,
+    typename new_executor_traits<Executor>::template future<
+      typename std::result_of<
+        Factory(typename new_executor_traits<Executor>::shape_type)
+      >::type
+    >,
     Future
   >,
   detail::type_list<Factories...>
@@ -71,14 +75,14 @@ struct use_single_agent_then_execute_with_nested_multi_agent_execute_with_shared
 using use_strategy_3 = use_single_agent_then_execute_with_nested_multi_agent_execute_with_shared_inits;
 
 
-template<class Container, class Executor, class Function, class Future, class... Factories>
+template<class Executor, class Function, class Factory, class Future, class... Factories>
 using select_multi_agent_then_execute_with_shared_inits_returning_user_specified_container_implementation =
   typename std::conditional<
-    //has_strategy_1<Executor, Container, Function, Future, Factories...>::value,
-    has_strategy_1_workaround_nvbug_1665745<Executor, Container, Function, Future, Factories...>::type::value,
+    //has_strategy_1<Executor, Function, Factory, Future, Factories...>::value,
+    has_strategy_1_workaround_nvbug_1665745<Executor, Function, Factory, Future, Factories...>::type::value,
     use_strategy_1,
     typename std::conditional<
-      has_strategy_2<Executor, Container, Function, Future, Factories...>::value,
+      has_strategy_2<Executor, Function, Factory, Future, Factories...>::value,
       use_strategy_2,
       use_strategy_3
     >::type
@@ -88,14 +92,16 @@ using select_multi_agent_then_execute_with_shared_inits_returning_user_specified
 
 // strategy 1
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class Future, class... Factories>
+template<class Executor, class Function, class Factory, class Future, class... Factories>
 __AGENCY_ANNOTATION
-typename new_executor_traits<Executor>::template future<Container>
+typename new_executor_traits<Executor>::template future<
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+>
   multi_agent_then_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_with_shared_inits_returning_user_specified_container_member_function,
-                                                                                Executor& ex, Function f, typename new_executor_traits<Executor>::shape_type shape, Future& fut,
+                                                                                Executor& ex, Function f, Factory result_factory, typename new_executor_traits<Executor>::shape_type shape, Future& fut,
                                                                                 Factories... shared_factories)
 {
-  return ex.template then_execute<Container>(f, shape, fut, shared_factories...);
+  return ex.then_execute(f, result_factory, shape, fut, shared_factories...);
 } // end multi_agent_then_execute_with_shared_inits_returning_user_specified_container()
 
 
@@ -117,15 +123,18 @@ struct strategy_2_functor
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class Future, class... Factories>
+template<class Executor, class Function, class Factory, class Future, class... Factories>
 __AGENCY_ANNOTATION
-typename new_executor_traits<Executor>::template future<Container>
+typename new_executor_traits<Executor>::template future<
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+>
   multi_agent_then_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_when_all_execute_and_select_with_shared_inits_member_function,
-                                                                                Executor& ex, Function f, typename new_executor_traits<Executor>::shape_type shape, Future& fut, Factories... shared_factories)
+                                                                                Executor& ex, Function f, Factory result_factory, typename new_executor_traits<Executor>::shape_type shape, Future& fut, Factories... shared_factories)
 {
   using traits = new_executor_traits<Executor>;
+  using container_type = typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type;
 
-  auto results = traits::template make_ready_future<Container>(ex, shape);
+  auto results = traits::template make_ready_future<container_type>(ex, result_factory(shape));
 
   auto results_and_fut = detail::make_tuple(std::move(results), std::move(fut));
 
@@ -135,11 +144,12 @@ typename new_executor_traits<Executor>::template future<Container>
 
 // strategy 3
 
-template<class Container, class Executor, class Function, class T, class Tuple>
+template<class Executor, class Function, class Factory, class T, class Tuple>
 struct strategy_3_functor
 {
   Executor& ex;
   mutable Function f;
+  mutable Factory result_factory;
   typename new_executor_traits<Executor>::shape_type shape;
   Tuple shared_inits;
 
@@ -161,24 +171,27 @@ struct strategy_3_functor
 
   template<size_t... Indices>
   __AGENCY_ANNOTATION
-  Container impl(detail::index_sequence<Indices...>, T& val) const
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+    impl(detail::index_sequence<Indices...>, T& val) const
   {
-    return new_executor_traits<Executor>::template execute<Container>(ex, inner_functor{f,val}, shape, std::get<Indices>(shared_inits)...);
+    return new_executor_traits<Executor>::execute(ex, inner_functor{f,val}, result_factory, shape, std::get<Indices>(shared_inits)...);
   }
 
   __AGENCY_ANNOTATION
-  Container operator()(T& val) const
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+    operator()(T& val) const
   {
     return impl(detail::make_index_sequence<std::tuple_size<Tuple>::value>(), val);
   }
 };
 
 
-template<class Container, class Executor, class Function, class Tuple>
-struct strategy_3_functor<Container,Executor,Function,void,Tuple>
+template<class Executor, class Function, class Factory, class Tuple>
+struct strategy_3_functor<Executor,Function,Factory,void,Tuple>
 {
   Executor& ex;
   mutable Function f;
+  mutable Factory result_factory;
   typename new_executor_traits<Executor>::shape_type shape;
   Tuple shared_factories;
 
@@ -199,46 +212,50 @@ struct strategy_3_functor<Container,Executor,Function,void,Tuple>
 
   template<size_t... Indices>
   __AGENCY_ANNOTATION
-  Container impl(detail::index_sequence<Indices...>) const
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+    impl(detail::index_sequence<Indices...>) const
   {
-    return new_executor_traits<Executor>::template execute<Container>(ex, inner_functor{f}, shape, detail::get<Indices>(shared_factories)...);
+    return new_executor_traits<Executor>::execute(ex, inner_functor{f}, result_factory, shape, detail::get<Indices>(shared_factories)...);
   }
 
   __AGENCY_ANNOTATION
-  Container operator()() const
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+    operator()() const
   {
     return impl(detail::make_index_sequence<std::tuple_size<Tuple>::value>());
   }
 };
 
 
-template<class Container, class Executor, class Function, class Future, class... Factories>
+template<class Executor, class Function, class Factory, class Future, class... Factories>
 __AGENCY_ANNOTATION
 strategy_3_functor<
-  Container,
   Executor,
   Function,
+  Factory,
   typename future_traits<Future>::value_type,
   detail::tuple<
     typename std::decay<Factories>::type...
   >
 >
-make_strategy_3_functor(Executor& ex, Function f, typename new_executor_traits<Executor>::shape_type shape, Future& fut, Factories... shared_factories)
+make_strategy_3_functor(Executor& ex, Function f, Factory result_factory, typename new_executor_traits<Executor>::shape_type shape, Future& fut, Factories... shared_factories)
 {
   auto factory_tuple = detail::make_tuple(shared_factories...);
 
-  return strategy_3_functor<Container,Executor,Function,typename future_traits<Future>::value_type,decltype(factory_tuple)>{ex, f, shape, factory_tuple};
+  return strategy_3_functor<Executor,Function,Factory,typename future_traits<Future>::value_type,decltype(factory_tuple)>{ex, f, result_factory, shape, factory_tuple};
 }
 
 
-template<class Container, class Executor, class Function, class Future, class... Factories>
+template<class Executor, class Function, class Factory, class Future, class... Factories>
 __AGENCY_ANNOTATION
-typename new_executor_traits<Executor>::template future<Container>
+typename new_executor_traits<Executor>::template future<
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+>
   multi_agent_then_execute_with_shared_inits_returning_user_specified_container(use_single_agent_then_execute_with_nested_multi_agent_execute_with_shared_inits,
-                                                                                Executor& ex, Function f, typename new_executor_traits<Executor>::shape_type shape, Future& fut,
+                                                                                Executor& ex, Function f, Factory result_factory, typename new_executor_traits<Executor>::shape_type shape, Future& fut,
                                                                                 Factories... shared_factories)
 {
-  auto g = make_strategy_3_functor<Container>(ex, f, shape, fut, shared_factories...);
+  auto g = make_strategy_3_functor(ex, f, result_factory, shape, fut, shared_factories...);
 
   return new_executor_traits<Executor>::then_execute(ex, g, fut);
 } // end multi_agent_then_execute_with_shared_inits_returning_user_specified_container()
@@ -250,17 +267,19 @@ typename new_executor_traits<Executor>::template future<Container>
 
 
 template<class Executor>
-  template<class Container, class Function, class Future,
+  template<class Function, class Factory, class Future,
            class... Factories,
            class Enable1,
            class Enable2,
-           class Enable3,
-           class Enable4
+           class Enable3
            >
-typename new_executor_traits<Executor>::template future<Container>
+typename new_executor_traits<Executor>::template future<
+  typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+>
   new_executor_traits<Executor>
     ::then_execute(typename new_executor_traits<Executor>::executor_type& ex,
                    Function f,
+                   Factory result_factory,
                    typename new_executor_traits<Executor>::shape_type shape,
                    Future& fut,
                    Factories... shared_factories)
@@ -268,14 +287,14 @@ typename new_executor_traits<Executor>::template future<Container>
   namespace ns = detail::new_executor_traits_detail::multi_agent_then_execute_with_shared_inits_returning_user_specified_container_implementation_strategies;
 
   using implementation_strategy = ns::select_multi_agent_then_execute_with_shared_inits_returning_user_specified_container_implementation<
-    Container,
     Executor,
     Function,
+    Factory,
     Future,
     Factories...
   >;
 
-  return ns::multi_agent_then_execute_with_shared_inits_returning_user_specified_container<Container>(implementation_strategy(), ex, f, shape, fut, shared_factories...);
+  return ns::multi_agent_then_execute_with_shared_inits_returning_user_specified_container(implementation_strategy(), ex, f, result_factory, shape, fut, shared_factories...);
 } // end new_executor_traits::then_execute()
